Nguyen

+++ src/main/java/com/example/do_an_tot_nghiep/service/IDashoardService.java
@@ -13,5 +13,5 @@ public interface IDashoardService {
 
     Map<String, Long> getOrderStatusCounts();
 
-    Long getPendingOrdersCount();
+    BigDecimal getPendingOrdersCount();
 }
